refactor into unitexpr, macros and coulomb subprojects
